requirements.txt
requests==2.31.0
python-dotenv==1.0.0
supabase==1.2.0
postgrest==0.11.0
python-jose==3.3.0
pydantic==2.6.1
pydantic-settings==2.2.1 